MCR-3441 some more code review changes
