+++ mycore-orcid2/src/main/resources/components/orcid2/config/mycore.properties
@@ -107,4 +107,4 @@ MCR.ORCID2.Object.TrustedIdentifierTypes=
 MCR.ORCID2.Mods.DateIssued.XPath=(mods:originInfo/mods:dateIssued[@encoding='w3cdtf'][@keyDate='yes'],descendant::mods:relatedItem[@type='host'][not(ancestor::mods:relatedItem[not(@type='host')])]/mods:originInfo/mods:dateIssued[@encoding='w3cdtf'][@keyDate='yes'],mods:originInfo/mods:dateIssued[@encoding='w3cdtf'],descendant::mods:relatedItem[@type='host'][not(ancestor::mods:relatedItem[not(@type='host')])]/mods:originInfo/mods:dateIssued[@encoding='w3cdtf'])[1]
 
 # Implementation for accessing all ORCIDs of a user
-MCR.ORCID2.AttributeHandler.Class=org.mycore.orcid2.user.MCRORCIDIDAttributeHandlerUserAttributeImpl
+MCR.ORCID2.AttributeHandler.Class=org.mycore.orcid2.user.MCRUserAttributeORCIDIDAttributeHandler
